examples/10/requirements.txt
# Multi-Agent Verification Loop System
# Python 3.8+ required

# Core dependencies
lancedb>=0.3.0
numpy>=1.21.0
requests>=2.28.0

# Optional: YAML config support
PyYAML>=6.0

# Optional: Better JSON handling
ujson>=5.7.0

# Optional: Progress bars
tqdm>=4.65.0

# Development dependencies (optional)
pytest>=7.2.0
pylint>=2.16.0
black>=23.1.0

# Note: LM Studio should be installed separately
# Download from: https://lmstudio.ai/
